refactor: update metrics module
