+++ src/metrics/policy_evaluations_total.rs
@@ -1,7 +1,7 @@
 use lazy_static::lazy_static;
 use opentelemetry::{metrics::Counter, KeyValue};
 
-use super::PolicyEvaluationMetric;
+use crate::metrics::PolicyEvaluationMetric;
 
 lazy_static! {
     static ref POLICY_EVALUATIONS_TOTAL: Counter<u64> =
